add function, 1-array_iterator

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,12 +11,12 @@
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
 
-int i;
+long unsigned int i;
 
 if (array && size && action)
 {
 
-for (i = 0; i < size ; i++)
+for (i = 0 ; i < size ; i++)
 action(array[i]);
 
 }
